Move `KeyPair` class to `Testnet.Runtime`
